<refactor> (VIS-91): Separação da Listagem de Fazendas em um novo componente dedicado
